dinnerparty/backend/requirements.txt
django==4.0.5
djangorestframework==3.13.1
django-debug-toolbar==3.4.0
django-cors-headers==3.13.0
PyJWT==1.7.1

pymongo==3.12.3
dnspython==2.2.1
pymongo[srv]==3.12.3
sqlparse==0.2.4

requests==2.28.0

djongo==1.3.6

django-rest-swagger==2.2.0
coreapi==2.3.3
pyyaml==6.0
drf_yasg==1.20.0

Pillow==9.1.1

psycopg2-binary==2.9.3

django-iso8601==1.1.0
uuid==1.30
numpy==1.23.1
